[Feature] Add QubeClient Facade for Unified API and Event Handling - Remove duplication between REST API and event logic by centralizing shared functionality.
Task:[T-6400]
