+++ qube/rest/tests/test_end_answering.py
@@ -15,7 +15,7 @@ from qube.rest.exceptions import (
     NotAuthorized,
     NotFound,
 )
-from qube.rest.types import Answering
+from qube.types import Answering
 
 
 @patch.object(RestClient, "put_request")
