+++ qube/rest/tests/test_list_queues.py
@@ -4,7 +4,7 @@ from unittest.mock import call, patch
 
 from qube.rest.clients import RestClient
 from qube.rest.exceptions import BadRequest, Forbidden, NotAuthorized, NotFound
-from qube.rest.types import Queue
+from qube.types import Queue
 
 
 @patch.object(RestClient, "get_request")
